Update and rename MockRiskOracle.sol to RiskOracle.sol
Renamed to RiskOracle as I am using the mock ReferralStorage.sol contract as an example, but can change this back if preferred.

Also updated mock RiskOracle contract to remove input params from the constructor. For initialAuthorizedSenders it's using msg.sender and for initialParameterTypes it's using a hardcoded list of all the current market parameter types.