Update ConfigSyncer.sol
Added allowedBaseKeys mapping, _initAllowedBaseKeys function and _validateKey function similar to how it is done in Config.sol (except _validateKey does not validating roles in this contract). This makes it so only parameters with specific baseKeys can be updated via ConfigSyncer and prevents updates for parameters with other baseKeys that aren't explicitly whitelisted.