Update MockRiskOracle.sol
Updated using final Chaos Labs RiskOracle contract. 

Previously, the constructor did not take inputs and instead had authorizedSenders[msg.sender] = true and the initialUpdateTypes[] hardcoded to include 50 parameters retrieved from updateMarketConfigUtils.ts but now the initialSenders and initialUpdateTypes are included as constructor input parameters and passed to the contract from the deployMockRiskOracle.ts script on deployment. 

The initialUpdateTypes has also been updated to align with the allowedBaseKeys included in the ConfigSyncer contract, plus maxPnlFactorForTradersLongs which is used in the ConfigSyncer.ts test script.

Please note that this updated MockRiskOracle contract is different from the final Chaos Labs RiskOracle contract in the following ways:

1. Changed to pragma solidity ^0.8.0 from pragma solidity ^0.8.25 because the latest version available in the repo is 0.8.18 and set to 0.8.0 to align with what is used for other contracts in the repo

2. Removed Ownable(msg.sender) from constructor as it appears the outdated version of the OpenZeppelin Ownable contract in the gmx-synthetics repo doesn't take this constructor input and just directly assigns Owner to msg.sender within the constructor

3. Changed the contract name from RiskOracle to MockRiskOracle