Update deployMockRiskOracle.ts
Updated to pass 'initialSenders' and 'initialUpdateTypes' as constructor param inputs to the updated MockRiskOracle contract on deployment.

The 'initialSenders' array is set to include just the deploying 'wallet' address used for tests in the ConfigSyncer.ts test script.

The 'initialUpdateTypes' array is set to include all parameters with base keys that equal true in the ConfigSyncer contract allowedBaseKeys plus maxPnlFactorForTradersLongs which is used for tests in the ConfigSyncer.ts test script. This should fix the two remediation findings from the Guardian audit (L-01 and L-02).